+++ drivers/drmem-drv-tplink/src/lib.rs
@@ -49,74 +49,7 @@ use tokio::{
 };
 use tracing::{debug, error, info, warn, Span};
 
-// This module defines the commands that can be sent to the TP-Link
-// device. It also configures the `serde` crate so these commands are
-// converted to the expected JSON layout.
-
-mod tplink_api {
-    use serde::{Deserialize, Serialize, Serializer};
-    use std::marker::PhantomData;
-
-    // Defines the internal value used by the `Active` command. Needs
-    // to convert to `{"state":value}`.
-
-    #[derive(Serialize, PartialEq, Debug)]
-    pub struct ActiveValue {
-        #[serde(rename = "state")]
-        pub value: u8,
-    }
-
-    // Defines the internal value used by the `Info` command. Needs
-    // to convert to `{}`.
-
-    #[derive(Serialize, PartialEq, Debug)]
-    pub struct InfoValue {
-	#[serde(skip)]
-	pub nothing: PhantomData<()>
-    }
-
-    // Defines the internal value used by the `Brightness` command. Needs
-    // to convert to `{"brightness":value}`.
-
-    #[derive(Serialize, PartialEq, Debug)]
-    pub struct BrightnessValue {
-        #[serde(rename = "brightness")]
-        pub value: u8,
-    }
-
-    #[derive(Serialize, PartialEq, Debug)]
-    pub enum Cmd {
-        #[serde(rename = "system")]
-        Active {
-            #[serde(rename = "set_relay_state")]
-            value: ActiveValue,
-        },
-
-        #[serde(rename = "system")]
-        Info {
-            #[serde(rename = "get_sysinfo")]
-            value: InfoValue,
-        },
-
-        #[serde(rename = "smartlife.iot.dimmer")]
-        Brightness {
-            #[serde(rename = "set_brightness")]
-            value: BrightnessValue,
-        },
-    }
-
-    pub fn active_cmd(v: u8) -> Cmd {
-	Cmd::Active { value: ActiveValue { value: v } }
-    }
-
-    pub fn brightness_cmd(v: u8) -> Cmd {
-	Cmd::Brightness { value: BrightnessValue { value: v } }
-    }
-
-    pub fn info_cmd() -> Cmd {
-	Cmd::Info { value: InfoValue { nothing: PhantomData } }
-    }
-}
+mod tplink_api;
 
 type Cmds = Vec<tplink_api::Cmd>;
 
@@ -187,10 +120,7 @@ impl Instance {
         // set the brightness and then turn on the dimmer.
 
         if v > 0.0 {
-            vec![
-                brightness_cmd(v as u8),
-                active_cmd(1),
-            ]
+            vec![brightness_cmd(v as u8), active_cmd(1)]
         } else {
             vec![active_cmd(0)]
         }
@@ -392,63 +322,26 @@ impl driver::API for Instance {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        tplink_api::{active_cmd, brightness_cmd, info_cmd},
-        Instance
-    };
-    use serde_json;
-
-    #[test]
-    fn test_cmds() {
-        assert_eq!(
-            serde_json::to_string(&active_cmd(1)).unwrap(),
-            "{\"system\":{\"set_relay_state\":{\"state\":1}}}"
-        );
-        assert_eq!(
-            serde_json::to_string(&info_cmd()).unwrap(),
-            "{\"system\":{\"get_sysinfo\":{}}}"
-        );
-        assert_eq!(
-            serde_json::to_string(&brightness_cmd(0)).unwrap(),
-            "{\"smartlife.iot.dimmer\":{\"set_brightness\":{\"brightness\":0}}}"
-        );
-        assert_eq!(
-            serde_json::to_string(&brightness_cmd(50)).unwrap(),
-            "{\"smartlife.iot.dimmer\":{\"set_brightness\":{\"brightness\":50}}}"
-        );
-        assert_eq!(
-            serde_json::to_string(&brightness_cmd(100)).unwrap(),
-            "{\"smartlife.iot.dimmer\":{\"set_brightness\":{\"brightness\":100}}}"
-        );
-    }
+mod test {
+    use super::{tplink_api, Instance};
 
     #[test]
     fn test_brightness_commands() {
         assert_eq!(
             Instance::set_brightness_cmd(0.0),
-            vec![active_cmd(0)]
+            vec![tplink_api::active_cmd(0)]
         );
         assert_eq!(
             Instance::set_brightness_cmd(1.0),
-            vec![
-                brightness_cmd(1),
-                active_cmd(1)
-            ]
+            vec![tplink_api::brightness_cmd(1), tplink_api::active_cmd(1)]
         );
         assert_eq!(
             Instance::set_brightness_cmd(50.0),
-            vec![
-                brightness_cmd(50),
-                active_cmd(1)
-            ]
+            vec![tplink_api::brightness_cmd(50), tplink_api::active_cmd(1)]
         );
         assert_eq!(
             Instance::set_brightness_cmd(100.0),
-            vec![
-                brightness_cmd(100),
-                active_cmd(1)
-            ]
+            vec![tplink_api::brightness_cmd(100), tplink_api::active_cmd(1)]
         );
     }
 }
